fix lof

+++ src/security/ecdh_create_key.cc
@@ -52,7 +52,7 @@ int EcdhCreateKey::Init() {
 int EcdhCreateKey::CreateKey(const PublicKey& peer_pubkey, std::string& sec_key) {
     auto secret_len = (field_size_ + 7) / 8;
     sec_key.resize(secret_len, 0);
-    if (secret_len != 32 || sec_key.size() != secret_len) {
+    if (secret_len != 32 || (int32_t)sec_key.size() != secret_len) {
         CRYPTO_ERROR("secret_len error: %d, str size: %d!", secret_len, sec_key.size());
         return kSecurityError;
     }
